选通道改成两级路由，外加 Lenovo 和 Acer 两条后端
先前"这台机器走哪条通道"是一串按顺序试出来的：候选表里 add 的次序就是答案，
类别只写在注释里。现在是查表：

一级路由按平台 —— Windows 一张表、Linux 一张表，表里每行只说"这是哪一类通道、
命中之后用哪个后端"，这一级不碰硬件。二级路由按识别依据 —— 厂商接口的 GUID
在不在、固件指纹认不认。各家的识别依据互不重叠，所以一个类别里最多命中一条，
同类别内谁先谁后不影响答案。平台是表的键而不是表里的一列：作为一列的话，
每行还要再声明一次"我只在 Windows 上能造出来"，那句声明和真正的平台判断是
两处，会漂移 —— 现在编译器也看得见这个边界。

**厂商接口一律按 GUID 找，不按 Windows 上的类名。** GUID 是固件里写死的，
上游实现也都用它指认；类名只是本地别名，各机型和驱动版本可能不一样，
按类名找等于在猜一个我们没有机器去核对的名字。方法同理按 MOF 里的编号找。
GUID 到类名的对照表整个进程只建一次 —— 建一次要过一遍这个 scope 的类定义，
本机 892 个类、1.5 秒，每条通道各扫一遍是用户能直接感觉到的慢。

新增两条后端（协议事实取自公开实现，只取事实不抄代码）：

- **Lenovo Legion**：读转速和开关强冷在两个不同的类上（LENOVO_FAN_METHOD 与
  LENOVO_GAMEZONE_DATA），所以它有两条 WMI 通道。
- **Acer Predator / Nitro**：这几条里唯一**给得了任意占空比**的厂商通道，
  它的接口直接收百分比。转速反而读不到（要走另一条加位掩码的路，各世代掩码
  不一样），所以如实报 null，不报一个可能是别的字段的数字。

两条都没有在对应机器上实测过，README 里如实写明。

本机回归只做只读检查（describe / read），没有去开关风扇：四条厂商通道都如实
不命中，仍然选中同方公版那条，风扇在固件自动模式。

+++ src/FanControlCore/Protocols/AsusWmiFanBackend.cs
@@ -174,7 +174,7 @@ public sealed class AsusWmiFanBackend : IFanBackend
     }
 
     /// <summary>读一项状态。固件说这一项不存在时返回 null，不返回 0。</summary>
-    private uint? ReadStatus(uint device)
+    private ulong? ReadStatus(uint device)
         => wmi.InvokeOrdered(ReadMethod, device) is { } raw && (raw & PresenceBit) != 0
             ? raw & DataMask
             : null;
